Added a new target to build a static library for iOS use. Fixed example.
git-svn-id: http://svn.mulle-kybernetik.com/OCMock/trunk@61 12620d81-68e2-0310-ae40-e3f550779089

+++ Source/OCMockRecorderTests.m
@@ -60,7 +60,7 @@
 	[recorder andReturn:@"foo"];
 	handlerList = [recorder invocationHandlers];
 	
-	STAssertEquals(1u, [handlerList count], @"Should have added one handler.");
+	STAssertEquals((NSUInteger)1, [handlerList count], @"Should have added one handler.");
 	STAssertEqualObjects([OCMReturnValueProvider class], [[handlerList objectAtIndex:0] class], @"Should have added correct handler.");
 }
 
@@ -73,7 +73,7 @@
 	[recorder andThrow:[NSException exceptionWithName:@"TestException" reason:@"A reason" userInfo:nil]];
 	handlerList = [recorder invocationHandlers];
 
-	STAssertEquals(1u, [handlerList count], @"Should have added one handler.");
+	STAssertEquals((NSUInteger)1, [handlerList count], @"Should have added one handler.");
 	STAssertEqualObjects([OCMExceptionReturnValueProvider class], [[handlerList objectAtIndex:0] class], @"Should have added correct handler.");
 	
 }
